backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]
python-multipart
python-keycloak
requests
pydantic>=1.8.2
python-dotenv
pymupdf
paddleocr
psycopg2-binary
minio
numpy
sentence-transformers
langchain
chromadb
python-docx
PyPDF2
passlib[bcrypt]
redis
sqlalchemy